Refactor jars page route and add Toaster to layout
Moved the jars/[jarId]/page.tsx and loading.tsx files to the new (pages) directory structure. Updated the JarPage component to use awaited params per Next.js 15 conventions and added error handling for unauthorized access. Also, added the Toaster component from 'sonner' to the root layout for global notifications.
